+++ ingenialink/ethercat/register.py
@@ -1,0 +1,5 @@
+from ingenialink.canopen.register import CanopenRegister
+
+
+class EthercatRegister(CanopenRegister):
+    pass